split entity parser tests by entity kind
